fixed schema - still no index but is working

+++ db/setup/schema.sql
@@ -7,7 +7,7 @@ CREATE TABLE customers(
   state text,
   postal_code text,
   phone text,
-  account_credit integer,
+  account_credit float,
   registered_at date
 );
